+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## :mouse: v0.5.0
+
+- ### :wrench: Maintenance
+
+  Clean out feature names:
+  - `pi3`: is now the feature to use the Raspberry Pi3 peripheral addresses
+  - `pi4_low`: use the Raspberry Pi4 Low-Peri mode peripheral addresses
+  - `pi4_high`: use the Raspberry Pi4 High-Peri mode peripheral addresses
+
+- ### :bulb: Features
+
+  Adding support for Raspberry Pi4 interrupt handling using the legacy interrupt controller.
+  
 ## :dog: v0.4.2
 
 - ### :wrench: Maintenance